Revert "[async-client] a taste of what it means to unblock the delegate"
This reverts commit 4d6c10a6956bb9a81144a61ebb6bcab3aedb840e.

This shows there is no sense in unblocking individual functions,
too much complexity and everything has to be assumed to be blocking
anyway except for receive_pack(…) which may only be unblocked if
if non-blocking IO is used to write a file directly. But nobody
ever does that and there is no non-blocking implementation yet.

There could be, of course, which would allow the operation to be
more easily aborted by stopping to 'iterate', but cancellation
can currently always be implemented by closing the connection
to make subsequent reads fail.