[ref #152] sketch API for namespaces
It's a little more elaborate in the way we integrate namespaces into
edits as it's less error prone and more explicit that way.

Only server side repositories will even see (or use) this feature,
but at least it requires no special knowledge except for knowing
that you use namespaces and applying them for listings and for edits.