[repository #165] FAIL Look into `owned_ref` crate
In an attempt to get around the GAT problem which prevents the kind of
generalization I need here I found the promising `owned_ref` crate's
ToHandle trait. It works around gats and provides the necessary unsafe
code for RefCell already, but not for Mutexes yet, see

https://docs.rs/owning_ref/0.4.1/src/owning_ref/lib.rs.html#1161-1163

Right now this would split the world in half as the Arc versions of
Access couldn't implement ToHandle, and working around would mean
to have some sort of type based condition in there that gets baked
into the returned Objects. Maybe, at least.

Note that OwnedRef types themselves carry lifetimes too which means
we can't use them as return types in a trait method.