treat clippy as generated statistical section… (#198)
…which comes with the benefit of not needing to be parsed back at all.
We will just treat it as marker when parsing, knowing that it will
be replaced with the same (or updated) information during merge.

This allows to create more elaborate statistics without adding
effort to parse this back - this effort is spent exlusively
where it matters, with the git-conventional sections that
are in fact user-editable.