fix: don't put more objects into the pack cache than needed (#198)
Previously when accessing a packed object, it would store the base
object into the pack cache (if it wasn't retrieved from there)
which is great if that operation is free.

Since it isn't, it's better not to stress the cache with puts
and trash more objects than necessary.

Now only the last decompressed object will be put into the LRU cache.