Respect user selection when re-adding crates for manifest change (#222)
When a user-selected crate was initially discarded as it had no change,
it may still be picked up again in case it is affected by breaking
changes.

When that happens, it would previously never partake in the release,
despite the breaking change, but will do now if the user provided it.